# Conflicts:
#	market-web-war/src/main/webapp/WEB-INF/views/vip-buy/vip-buy.jsp
#	market-web-war/src/main/webapp/static/dist/js/findPassword.js
#	market-web-war/src/main/webapp/static/dist/js/regLogin.js

+++ market-web-war/src/main/webapp/static/src/assets/js/order/order.js
@@ -157,7 +157,7 @@
             return;
         }
         orderFlag = true;
-        var amount = parseInt(price || '200');
+        var amount = parseFloat(price || '200');
         var openId = cookie.getCookieValue('openId');
         util.ajaxFun(interfaceUrl.payOrder, 'POST', {
             orderNo: orderNo,
